Remove config.local option
