+++ api/lib/connection-pool.ts
@@ -87,7 +87,7 @@ export default class ConnectionPool extends Map<number | string, ConnectionClien
 
         return new Promise((resolve, reject) => {
             stream.on('data', async (conn: InferSelectModel<typeof Connection>) => {
-                if (conn.enabled && !this.config.local) {
+                if (conn.enabled) {
                     conns.push(this.add(new MachineConnConfig(this.config, conn)));
                 }
             }).on('error', (err) => {
